New version, man page, exciting changes!!
Bugs:
1. issue #200 - forgot to add all variants for -p, now works with --partition-full
and --partitions-full
2. issue #199 - another one, forgot to add --disk to -D for long version. Thanks
adrian15 for both of these, he was testing something and discovered these were
missing.
3. Issue #187 an issue with RAID syntax not being handled in a certain case,
thanks EnochTheWise for following through on this one. This turned out to be
a bad copy paste, a test pattern did not match the match pattern.

Fixes:
1. Fixed some docs typos.
2. Issue #188 fixed protections and filters for some glxinfo output handlers.
3. Issue #195, for Elbrus bit detection.
4. Added filter to cpu data, was not skipping if arm, so Model string
was treated numerically.

Enhancements:
1. Added rescatux to Debian system base detections. This closes issue #202, again
from adrian15, thanks.
2. For cpu architecture, updated for latest AMD ryzen and other families, like
Zen 3, which is just coming out re available data. Also latest Intel, which are
trickier to ID right now, but I think I got the latest ones right,
That's things like coffee lake, amber lake, comet lake, etc.
3. Huge one, full (hopefully out of the box) Russian Elbrus CPU support. Thanks
to the alt-linux and the others who helped provide data and feedback to get support.
Note that this was also part of correcting 64 bit detection for e2k type, which
is how Elbrus IDs internally. See issue #197 which I've left open for the time
being for more information on this CPU and how it's now handled by inxi.
Note all available data should now work for Elbrus, including physical cpu/core
counts etc. Elbrus do not show flag information, nor do they use min/max speed,
so that data isn't available, but everything else seems to work well.
4. Eternal disk vendors. Thanks linux lite hardware database, you continue to
help make the disk vendor feature work by supplying every known vendor ever seen.
5. To close debian bug report https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=942194
Note that the fix is simply to give the user the option to disable this behavior
with the new --no-sudo and NO_SUDO configuration file options. This issue should
never have been filed as a bug since even the poster admitted it was a wishlist
item, but because of how debian bug tracker works, it's hard to get rid of
invalid bugs. Note that this is the internal use of sudo for hddtemp and file,
not starting inxi with sudo, so using this option or configuration item just
removes sudo from the command. Note that because the user did not do as
requested, and never actually filed a github wishlist issue, and since his
request was vague and basically pointless, the fix is just to let you switch
off sudo, that's all.