Bugs: 1. Fixed undefined error that could happen, in rare cases, in hdd_temp logic.
Fixes:
1. Fixed Elbrus cpu nazming, model 9 is 8CV, not 8CB (Cyrillic error)
2. Preventitive, was not using '-' quite correctly in all regex ranges.
3. Had wrong desktop string listed in Unity
4. Reordered Family/Drive model in usb drive reports, it's to make it
more obvious what is what.
5. Adjusted indexing of splits to get better results in corner cases.
6. Fixed some numbering issues.
7. Added trimming n1 from nvme0 type names for nvme, this corrects some
issues users were having.
8. Fixed a division by 0 error in smartctl data grabber.
9. Fixed a Perl issue, didn't realize perl treats 000 as a string, not 0.
10. Another Perl fix, int() only wants to get numeric values sent to it,
I'd assumed a different behavior, non numerics get converted to 0, but that's
not how Perl sees things. Things like this, by the way, are why Perl is so
absurdly fast.

Enhancements:
1. More disk vendors. The list will never be complete!! We have found eternal
churn!! Thanks to linux lite hardware database as always.
2. Big one!!! Now inxi uses smartctl data, if installed, for getting advanced
drive information (with -a). See man and help for details. Will show failing drives,
etc. Lots of info can be available, but sometimes data is not in smartctl db,
so inxi can't find it, that's not an inxi bug, it's just how it is.
3. Made hours on more human readable, into days/hours, for -a smartctl disk
report.
4. Added $test[12] for smartctl data printout, and $test[13] for disk array print out.
Note that advanced debugger outputs can change or vary depending on what is being
worked on so don't in general rely on these always being around. But they do
tend to say stuck in place once I add them.
5. Added some nvme stuff, spare reserve, if you need it, you'll appreciate it,
if not, you'll never know it's there.
6. By request from some forum issue thread: made --host only be shown onif not
--filter or not --host. This makes -z remove hostname, but retains ability to
do absolute overrides. Hostname should have always been filtered out like that,
it was an oversight. I think that was Manjaro who asked that, but I forget.
Note that this change, as usual, will not alter expected behaviors if users
have config item for hostname set.
7. Added support for picom compositor, thanks user codebling for that, I think
that's compiz fork, the real branch that is that is being developed.