Bug fixes!!! New Features!! Why wait!!!
Bugs:
1. Issue #220 on github: inxi misidentified XFCE as Gnome. This was a kind of core
issue, and pointed to some logic that needed updating, and some inadequate
assumptions made, and some too loose cascade of tests. Hopefully now xfce will
almost never get misidentified, and the other primary desktops ID'ed either from
$ENV or from xrop -root will be slightly more accurately identified as well.

Note that this fix creates a possibility for obscure misconfigured desktops to
be ID'ed wrong, but in this case, that will be technically a bug for them, but
with the new fixes, that situation will be cleaner to handle internally in the
desktop ID logic.

Also tightened the final Gnome fallback detection to not trigger a possible
false positive, it was testing for ^_GNOME but that is not adequate, because
some gnome programs will trigger these values in xprop -root even if GNOME
is not running. Should be safer now, hopefully no new bugs will be triggered
by these changes.

Fixes:
1. Missed an indentation level for -y1, gcc alt should have been indented in
one more level, now it is.

2. In disk vendors/family, didn't clean items starting with '/', this is
now corrected. Yes, some do, don't ask me why. Might be cases like:
Crucial/Micron maybe, where the first ID is grabbed, not sure.

Enhancements:
1. New Disk vendors, vendor IDs!!! The list never ends!!! We've finally found
infinity, and it is the unceasing wave of tiny and not so tiny disks and their
Ids.

2. New feature: for -Aa, -Na/-na/-ia, -Ga, now will add the modules the kernel
could support if they were available on the Device-x lines of those items.
This was made an -a option because it really makes no sense, if it's a regular
option, users might think that for example an nvidia card had a nouveua driver
when it didn't, when in fact, all the kernel is saying is that it knows those
listed modules 'couid' be used or present. This corresponds to the Display:
item in -Ga, that lists 'alternate:' drivers that Xorg knows about that could
likewise be used, if they were on the system.

In other words these are --admin options because otherwise users might get confused,
so this is one where you want to know the man explanation before you ask for it.

It is useful however if you're not sure what your choices are for kernel modules.

When the alternate driver is the same as the active driver, or if none is found,
it does not show the alternate: item to avoid spamming.