New features, new changes, new bug fixes!!! Excitement!!! Thrills!!!
Bugs:
1. Forgot to set get Shell logic in inxi short form, oops, so Shell remained blank,
only inxi short, which I rarely use so I didn't notice.

2. Failed to test pacman-g2 for packages, had wrong query argument, so it failed.
Also failed to test for null data, so showed errors for packages as well. Both
fixed.

3. A big bug, subtle, and also at the same time, an enhancement, it turns out NVME
drives do NOT follow the age old /proc/partitions logic where if the minor number is
divisible by 16 or has remainder 8 when divided by 16, it's a primary drive, not
a partition. nvme drives use a random numbering when > 1 nvme drives are present, and
the old tests would fail for all nvme drivers more than the first one, which led
to wrong disk size totals. Thanks gardotd426 who took the time to help figure this
out in issue #223 - fix is to not do that test for nvme drives, or rather, to add
a last fail test for nvme primary nvme[0-9]n[0-9] drive detections, not the minor
number.

Fixes:
1. Corrected indentation for block sizes, children were not indented.

2. Updated some older inxi-perl/docs pages, why not, once in a while?

3. Kernel 5.8 introduces a changed syntax to gcc string location, this has been
corrected, and the kernel gcc version now shows correctly for the previous
syntax and the new one. Hopefully they do not change it again, sigh...

4. Removed string 'hwmon' sensors from gpu, those are not gpu sensors, and
are also usually not board/cpu sensors, but things like ath10, iwl, etc,
network, or disk sensors, etc. In some cases hwmon sensor data would appear

Enhancements:
1. Big sensors refactor, now inxi supports two new sensors options:
--sensors-exclude - which allows you to exclude any primary sensor type[s]. Note that
in the refactored logic, and in the old logic, gpu sensors were already excluded.
Now other hardware specific sensors like network are excluded as well.

--sensors-use - use ONLY list of supplied sensor IDs, which have to match the
syntax you see in lm-sensors sensors output.

Both accept comma separated list of sensors, 1 or more, no spaces.

The refactor however is more far reaching, now inxi stores and structures data
not as a long line of sensors and data without differentiation, but by sensor array/chip
ID, which is how  the exclude and use features can work, and how granular default
hardware sensor exclusions and uses can happen. This is now working in the gpu
sensors, and will in the future be extended to the newer 5.7/5.8 kernel disk
temperature sensors values, which will lead in some cases to being able to get
sensors data for disks without root or hddtemp. This is a complicated bit of logic,
and I don't have time to do it right now, but the data is now there and stored
and possible to use in the future.

To see sensors structures, use: inxi -s --dbg 18 and that will show the sensors data
and its structures, which makes debugger a lot easier for new features.

This issue was originally generated by what was in my view an invalid complaint
about some inxi sensors defaults, which led me to look more closely at sensors
logic, which is severely lacking. More work on sensors will happen in the future,
time, health, and energy permitting.

2. Added Watts, mem temp, for amdgpu sensors, as -sxxx option. More gpu sensor
data will be added as new data samples show what will be available for the
free modules like amdgpu, nouvean, and the intel graphics modules.

3. More disk vendors and IDs, as noted, the list never ends, and it hasn't ended,
so statement remains true. Thanks linux-lite hardware database.

Changes:

1. This has always bugged me since it was introduced, the primary cpu line starter
Topology: which was only technically accurate for its direct value, not its children,
and also, in -b, cpu short form was using the value as the key, which is a no-no,
I'd been meaning to fix that too, but finally realized if I just make the primary
CPU line key be 'Info:', which is short, yet non-ambiguous, it would solve both
problems.
To keep the -b cpu line as short as before, I removed the 'type:' and integraged
that value into the primary Info: string:
CPU:
  Info: 6-Core AMD Ryzen 5 2600 [MT MCP] speed: 2750 MHz min/max: 1550/3400 MHz

-b 3.1.05 and earlier:
CPU:
  6-Core: AMD Ryzen 5 2600 type: MT MCP speed: 1515 MHz min/max: 1550/3400 MHz

These resolve something that has irked me for quite a while, 'Topology:' didn't
fit, it was too geeky, and worst, it only applied to the value directly following
it, NOT to the rest of the CPU information. It also could not be shortened or
abbreviated since then it would have made no actual sense, like topo:, and the
same issue with value being used for key in -b, and wrong word for line starter
in -C would have existed. Besides, someone might think I was trying to make a
subtle reference to the great Jodorowsky film 'El Topo', which would be silly,
because that's art, and this is just some system specs that are reasonably
readable...

2. Was using opendns for WAN dig IP address, but apparently cysco bought that
company, and now I've noticed the old opendns dig queries were failing more and
more, so replaced that with akamai dig requests.
Also made the WAN IP fallback to HTTP IP method if dig failed. New option:
--no-http-wan and config item NO_HTTP_WAN with override --http-wan added to
let you switch off http wan IP requests if you want. Note that if dig fails,
you will get no wan ip address.
Updated/improved error messages to handle this more complex set of wan ip
options, so hopefully the error alert message will in most cases be right.

3. To future proof inxi, switched debugger upload location to ftp.smxi.org/incoming
from the old techpatterns.com/incoming. Updated man/help to remove those urls too.