Bug fixes, updates!!! Yes!! Why wait!!! Can't stay frozen forever!
Bugs:
1. Not an inxi bug, but a weird change in defaults for ubuntu GNOME ENV
variable values when running at least the gnome desktop, result to end
users appears to be a bug. This resolves issue #228
Note that so much weird non desktop data was put into those environmental
variables that inxi simply could make no sense of it. The fix was to make
the detections more robust, using regex instead of string compare, as well as
to at least try to strip out such corrupted data values, though that can never
be fully predictable.
As far as I know, this issue only hits ubuntu gnome desktops, I've never seen these
value corruptions on any other distro, or on any other ubuntu desktop, though
they may be there, but I'm not going to test all the ubuntu spins to find out.

I'm hoping the combination of logic fixes and junk data cleaning will handle
most future instances of these types of corruptions automatically.

Again, this only happens on relatively laste ubuntu gnomes as far as I know.

Fixes:
1. An oversight, added sshd to list of whitelisted start clients. This permits
expected output for: ssh <name@server> inxi -bay
that is, running inxi as an ssh command string. Should have done that a while ago,
but better late than never.
This corrects issue #227, or at least, has a better default, it worked fine before,
but required using --tty to reset to default terminal behavior. The problem is
that if inxi can't determine what it's running in, it defaults to thinking it's
in an IRC client, and switches to IRC color codes, among other changes.
But it was nice to get sshd covered automatically so users don't have to know
the --tty option.

Changes:
1. More disk vendors and vendor IDs!!! Yes, that's right, the list never ends!!