Bug fixes, new features!! Update now!! Or don't, it's up to you.
Bugs:
1. Let's call some of the android fixes and debugger failures bugs, why not?
Those are fixed. Note that many of these fixes will impact any system that is
ARM based, not just android.

Fixes:
1. Related to issue #226 which was a fine issue, fine tuned the debugger debuggers
to allow for smoother handling of /sys parse failures. Also added debugger filters
for common items that would make the /sys parser hang, oddly, most seem to be in
/sys/power for android devices.

2. Added some finetunings for possible mmcblk storage paths, in some cases, an
extra /block is added, which made inxi think mounted drives were unmounted. I've
never seen this extra /block except on mmcblk devices on android, but you never
know, it could be more widespread.

3. Also mainly related to android, but maybe other ARM devices, in some cases,
an errant 'timer' device was appearing as a cpu variant, which is wrong. That was
a corner case for sure, and part of the variant logic in fact uses timer values
to assign the actual cpu variants, but it was wrong in this case because it was
....-timer-mem, not ...-timer, which led to non-existent CPU variants showing.

4. Issue #236 by ChrisCheney pointed out that inxi had never updated its default
/proc/meminfo value to use the newer MemAvailable as default if present, which led
to incorrect memory used values showing up. That's because back in the old days,
we had to construct a synthetic Memory used from MemFree, buffers, cache, etc, but that
wasn't always right, since sometimes the cache actually isn't available, often is,
but not always.
https://github.com/torvalds/linux/commit/34e431b0ae398fc54ea69ff85ec700722c9da773
This commit on the kernel explains it pretty clearly.
Thanks Chris for bringing this to our attention.

5. Kind of more future-proofing, got rid of a bunchy of hard-coded strings internally
and switched those to use the row_defaults values, which is where string messages
are supposed to go. That was mostly in the initial program check messages on start-up,
but also a few other stray ones. Also consolidated them a bit to get rid of redundant
messages, and added more variable based messages, like for missing/permissions on
programs etc. The idea in general is that all the strings are contained in subs so
that in theory they could be swapped for other strings, eg, languages, but honestly,
I no longer see this as very likely to ever happen. But it's still nice to be
consistent internally and not get sloppy with english strings.

This also got rid of some largely redundant items in row_defaults, and expanded the
list of handled events, and of variable based events, so it shouldn't be as necessary
to add new row_defaults items for similar events.

Enhancements:
1. Debugger item to maybe try to find distro OEM, this was connected with issue #231
but the issue poster vanished, and didn't do the work required, so this one won't
happen until someone who cares [not me, that is] does the required work.
It's always funny to see how quickly people vanish when they have to do the actual
boring research that they want me to do for them, lol. Or maybe, sigh is more
appropriate than lol. But it is pretty much par for the course, sad to say.
Or maybe this was an OEM hoping to have someone do their corporate work for them
for free, who knows. Anyway, there's a certain category of items that I'm reasonably
happy to implement, but NOT if I have to do all the boring research work, so such
features being added will depend on the poster actually doing the boring work.

I've gotten burned on this a few times, cpu arch: for example, some guy said he'd
track that and provide updates, he never even made it to the first release, so I got
stuck doing that one forever after. But that one at least has some general value, so
that's ok more or less, but I definitely won't take on stuff that I really don't
personally care at all about unless the person requesting the feature does all the work
beforehand. The boring part, that is....

2. Related to issue #226, much improved android ID and many small android fixes for
machine data etc. Now uses /system/build.prop for some data, which is a nice source,
sadly, most modern android devices seem to be locked down, with both build.prop and
/sys locked down, which makes inxi unable to actually get any of that data, but if
your device either does not have these root only readable, or if you have an android
rooted phone, the android support will be more informative.

Hint: if you run inxi in termux on your non rooted android device, and it shows
you what android version you are using in System:... Distro: line, then your android
is not locked down. I have one such phone, android 7.1, but I cannot say how usual
or non usual this is. The poster of issue #226 for instance had to root his android
7 phone to get this data to display. So it seems to vary quite a bit.

Note that due to these file system lockdowns, in general, trying to do android arm
support remains largely a waste of time, but on some devices sometimes, you can now
get quite nice system info. As I noted in the issue, if I can't get the features to
work on a non rooted phone in my possession, I'm probably not going to try to do the
work because it's too hard to try to work on android issues without having the device
in front of you for testing and debugging. In this case, one of my phones did work, so
I did the work just to see where android is at now.

Android showed some slightly odd syntaxes for some devices, but those are now handled
where I got a dataset for them that revealed the changes required.

3. Also related to issue #226 for termux in android, will show -r info.
That's an apt based package manager, but termux puts the apt files somewhere else so
needed to change paths if those alternate paths existed for apt.

4. Added PARTFLAGS to debugger to see what knd of data that will yield, that's
a lsblk key/value pair.

5. Just because it's easy to do, added new -Ixxx item, wakeups: which is a
subset of Uptime, this will show how many times the system has been woken from
suspend since the last boot. If the system has never been suspended, shows 0.

6. Many more disk vendors and disk IDs. The list just never ends, possibly a
metaphor for something, the endless spinning of maya, who knows?

7. Added newest known ubuntu release, hirsute, to buntu ID logic. Might as well
catch them early, that will be 21.04.