Bug Fixes!!! Continuing internal refactor!!
This bug report came in right after 3.2.00 went out live, but I would never have
found it myself in testing so better found than not found!

Bugs:

1. A bug was introduced to dmidecode data handlers in 3.2.00 resulted in the
dmidecode data array basically eating itself up until errors appear. Quite difficult
to trigger, but babydr from Slackware forums figured it out, using -F --dmidecode
to force dmidecode use for all features that support it triggered thee bug always.
This was a result of the refactor, previously inxi had worked on copies of referenced
arrays, but in this case, it was working on the original array of arrays, subtle,
but obvious. This method was only used on dmidecode arrays.

2. A second bug was exposed almost by accident, for -M --dmidecode data, there was
a missing field and also a missing is set test on that field that led to an error
of using undefined value in string comparison. This was strictly speaking 2 bugs,
both very old, from 2.9 first rewrite, one failing to set/get the value, and the
other failing to test if the value was set before using it.

Fixes:

1. There were a few glitches in help menu and man page related to -L option, those
are corrected.

INTERNAL CODE CHANGES:
1. removed bug inducing splice use in some cases, and added parens to splice to make
it fit the new way of with perl builtins, when taking 2 or more arguments, use parens.

2. Found many more instances to add -> dereferencing operator. I have to say, not
doing that consistently made the code much harder to read, and created situations
where it's somewhat ambiguous what item belongs to what, with everything consistently
-> operator run, the code is more clear and obvious, and some of the hacks I'd added
because of the lack of clarity were also removed.

3. Removed explicit setting of hash references with null value, that was done out
of failure to use -> operators which clearly indicate to Perl and coder what is
happening, so those crutches were removed. Also got rid of unnecessary array
priming like: my @array = (); Some of these habits came from other languages,
but in Perl, declaring my @array means it's an array that is null, and you don't
need to do a further (). @array = () is obviously fine for resetting arrays in
loops or whatever, but not in the initial declaration.