New version, man page, bug fixes, changes, adjustments and cleanups!!!
Special thanks to mr. mazda for his ongoing suggestions, ideas, and observations.

Bugs:
1. In certain corner cases, it appears that lsusb has blank lines, which tripped
errors in inxi output when the usb parser was trying to access split keys that did
not exist. Added in check to make sure split actually resulted in expected data.

2. A red face bug, I'd left the output debugger switched on with json output, so
it was printing out the json data structure with Dumper, that's now switched off.
Hope this doesn't mess anyone up, but it would have mattered only if the person
was using:
--output json --output-type print
It did not effect xml output.

Fixes:
1. Got rid of extra level of -L data structure and output handler. Not visible
to users, but still irksome, so nice to get that fixed. Recursive structures are
confusing, lol, but this extra level was pointless, but to fix it required redoing
the logic a bit for both data generator and output feature.

2. Added in support for --display :0.0, previously it did not support the
.0 addition, but why not, if it works for people, good, if not, makes no difference.

3. There were some missing cases for LVM missing data messages, so the following
fixes were added:
 * In cases where lsblk is installed and user is non root, or lvs is not installed,
 but no lvm data is present, inxi now shows the expected 'Message: No LVM data found.'
 instead of the permissions or missing program error that showed before.
 If lsblk is not installed, and lvm is installed (or missing), with lvs not root
 readable, the permissiosn message (or missing program) will show since at that
 point, inxi has no way to know if there is lvm data or not.

 * Not an inxi, but rather an Arch Linux packaging bug, the maintainer of lvm
 has made lvs and vgs fail to return error number on non root start, which is
 a bug (pvs does return expected error return). Rather than wait for this bug
 to be fixed, inxi will just test if lvs and lsblk lvm data, it will show
 permissions message, otherwse the no lvm data message as expected.

I think these cover the last unhandled LVM cases I came across, so ideally, the
lvm data messages will be reasonably correct.

4. Some man page lintian fixes.

5. Changed usb data parser to use 'unless' instead of 'if' in tests since
it's easier to read unless positive tests are true than if negative or
negative etc.

Enhancements:
1. Since I see too often things like -F --no-host -z which is redundant, the
help and man now make it more clear that -z implies --no-host.

2. Even though it's not that pointful, I added in derived Arch Linux system
base like Ubuntu/Debian have. It's not that meaningful because unlike
Ubuntu/Debian, where you want to know what version the derived distro is
based on, Arch is rolling thus no versions, but I figured, why not, it's
easy to do, so might as well make the system base feature a bit more complete.

Note that the way I did this requires that the distro is ID'ed as its derived
distro nanme, not Arch Linux, that will vary depending on how they did their
os-release etc, or distro files, but that's not really an inxi issue, that's
up to them. From what I've been seeing, it looks like more of the derived
distros are being ID'ed in inxi as the derived name, so those should all work
fine. Note that seeing 'base:' requires -Sx.

3. More disk vendors!! More disk vendor IDs!!! I really dug into the stuff,
and refactored slightly the backend tools I use, so it's now a bit easier
to handle the data. Thanks linux-lite hardware database, as always, for
having users that really seemt to use every disk variant known to humanity.

Changes:
1. In -G, made FAILED: lower case, and also moved it to be after unloaded:
It was too easy to think that the loaded driver had failed. Also to make it
more explicit, made output like this, in other words, driver: is a container
for the possible children: loaded: unloaded: failed: alternate: which should be
easier to parse and read without mixing up what belongs to what.

driver: loaded: modesetting unloaded: nouvean,vesa alternate: nv
driver: loaded: amdgpu unloaded: vesa failed: ati

Note that if there is no unloaded: driver, failed: would still appear to come after
loaded:, but hopefully it's more clear now.

Basically what we found was that the presence of the uppercase FAILED: drew
the eye so much that it was sometimes not noted that it was a key: following
the driver: item, which itself because it did not list explicitly loaded:
was not as clear as it could have been. By making failed: the same as the
other key names visually, hopefully it will be less easy to think that the
loaded: driver failed:

In a sense, this is a legacy issue, because the original use of FAILED: was for
non free video drivers, to see when xorg had failed to load them, but over
more recent years, the most frequent thing I have been seeing is odd things
like failed: ati, when xorg tries to load the legacy ati driver when amdgpu
is being used.

2. Likewise, for RAID mdraid and zfs changed FAILED: to Failed:, again, to make it
more consistent with the other types.

3. In help menu and man page, removed legacy 'card(s)' in -A, -G, -N, and replaced
that with 'device(s)', which is the more accurate term, since the days when these
things were only addon cards are long behind us. I had not noticed that, but it
caught me eye and I realized it was a very deprecated and obsolete syntax, which
did not match the way inxi describes devices today.

4. It was pointed out how incoherent the naming of the item for setting wrap width,
--indent-min and config item INDENT_MIN were super confusing, since it was neither
indent or minimum, it was in fact wrap maximum, so the new options and config items
are --wrap-max and WRAP_MAX. Note that the legacy values will keep working, but
it was almost impossible in words to explain this option because the option text
was almost the exact opposite of what the option actually does. Redid the man
and help explanations to make the function of this option/config item more clear.

5. Made -J/--usb Hub-xx: to fit with other repeating device types in inxi output,
before Hub: was not numbered, but it struck me, it should be, like all the other
auto-incremented counter line starters, like ID-xx:, Device-x:, and so on.

6. Reorganized the main help menu to hopefully be more logical, now it shows the
primary output triggers, then after, the extra data items, -a, -x, -xx, -xxx,
separated by white space per type to make it easier to read. This also moved
the stuff that had been under the -x items back to where they should be, together
with the main output control options. For readability and usability, I think this
will help, the help menu is really long, so the more visual cues it has to make it
clear what each section is, the better I think. Previously -a was the first items,
then way further down was -x, -xx, and -xxx, then under those was -z, -Z, -y.