Bug fixes!! Fixes!!! Refactors!!! Edits!!!
Bugs:
1. Big bug, 3.2 appears to have introduced this bug, for disks, rotation and
partition scheme would never show, oops.

2. Tiny bug kept one specific smart value from ever showing, typo.

Fixes:
1. Accidentally followed Arch linux derived distro page, which claims KaOS as
arch derived, when of course it's not, it's its own distro, own toolchain, etc.
I kind of knew this but had forgotten, then I believed the Arch derived distro
page, oh well. Resulted in KaOS being listed with arch linux as system base
with -Sx. Arch should fix this, it's not like it's hard, just remove the distro
from the page.

2. Cleared up explanations for drivetemp vs hddtemp use, updated --recommends,
man, and help to hopefully make this clear. Debian will be dropping hddtemp,
which is not maintained, sometime in the coming years, sooner than later.
Note that users unfortunately have to manually enable drivetemp module unless
their distros enable it by default, but the man/recommands/help explain that.

3. Fixed smart indentation issues, that went along with code change 1, was
failing to indent one further level for failed/age values like it's supposed
to.

Enhancements:
1. Added /proc/device to debugger, that will help track block device main numbers

2. More disk vendors, more disk vendor IDs!!! As noted, the enternal flow flows
eternally, thanks linux-lite hardware database users!! and other inxi users,
whose outputs sometimes reveal a failure or two.

3. Added loaded kernel module tests to --recommends, this was mostly to let users
know that drivetemp is needed if you want non superuser fast drive temps, and
that this came along with kernels 5.6 or newer. Hopefully word will start drifting
out. Note that if inxi is using drivetemp values, drive temps will appear as
regular user with -Dx, and will be to 1 decimal place. hddtemp temps are
integers, and requires sudo to display the temps.

4. To handle issue #239  which I'd thought of trying off and on, but never did,
added option to -Dxxx to show SSD if a positive SSD ID was made to rotation:
So rotation will show either nothing, if no rotation or ssd data is detected,
the disk speed in rpm, or SSD if an SSD device. There may be corner cases where
this is wrong, but I don't have data for that, for example, if a disk is parked
and has zero rotation but is a HDD, not as SSD. I don't know what the data
looksl ike in that case. Note that if sudo inxi -Da is used, and smartctl is
installed, it should be right almost all the time, and with regular -Dxxx, it's
going to be right almost always, with a few corner cases. That slight
uncertainty is why I never implemented this before. Legacy drives also sometimes
did not report rotation speeds even when HDD, so those may create issues,
but inxi will only call it an SSD if it's an nvme, mmcblk device, both are
easy to ID as SSD, or if it meets certain conditions. It will not call a drive
an SSD if it was unable to meet those conditions.

INTERNAL CODE CHANGES:
1. Refactored the output logic for DiskData, that was messy, split it into a few
subs, and also refactored the way smartctl data was loaded and used, that's
much cleaner and easier to use now. Split the previous 1 big sub into:
totals_output(), drives_output(), and smart_output().
Also split out the smart field arrays into a separate sub, which loads
references to avoid creating new arrays and copying them all over when outputting
smart data. References are weird to work with directly but they are MUCH faster
to use, so I'm moving as much of the internal logic to use array raferences
instead of dereferenced arrays/hashes assigned to a new array, or hash.

2. Redid all the output modules and renamed them to be more consistent and
predictable, and redid the logic here and there to make the get() items be fairly
similar on all the data builder packages. Now as with the data subs, which
generally end in _data, now most of the output subs end with _output.

3. Roughly finished the process started in 3.2, got rid of redundant array loads,
changed:
@something = something_data();
push (@rows,@something);
to:
push (@rows,something_data());
which avoids creating an extra array, this also let me remove many arrays overall.

4. Missed a few hashes in machine data that were being passed directly, not as
references, to other subs, corrected that. I think I missed those because they
were %, so the search I did for @ in sub arg lists didn't catch the % hashes.